Menampilkan data yang akan diubah

+++ app/src/main/java/com/jesika0104/todolist/navigation/Screen.kt
@@ -6,6 +6,6 @@ sealed class Screen(val route: String) {
     data object Home: Screen("mainScreen")
     data object FormBaru: Screen("detailScreen")
     data object FormUbah: Screen("detailScreen/{$KEY_ID_TASK}") {
-        fun withId(id: Int) = "detailScreen/$id"
+        fun withId(id: Long) = "detailScreen/$id"
     }
 }
